ipsec-tests: Fix flaky TestUpsertIPSecKeyMissing
In case previous test failed, it did not clean up ipSecKeysGlobal
resulting in failing TestUpsertIPSecKeyMissing too.
Let's ensure that global ipsec key is cleaned up.

Fixes: #32933
